+++ lib/redis-sliding-window-rate-limiter.d.ts
@@ -14,14 +14,14 @@ export class RedisSlidingWindowRateLimiter extends EventEmitter implements Slidi
 
   constructor (options?: RedisSlidingWindowRateLimiterOptions)
 
-  reserve (key: string, limit: number): Promise<number>
-  reserve (key: string, limit: number, callback: (error: Error | null, ts: number) => void): void
-
   check (key: string, limit: number): Promise<number>
   check (key: string, limit: number, callback: (error: Error | null, usage: number) => void): void
 
-  cancel (key: string, limit: number, ts: number): Promise<number>
-  cancel (key: string, limit: number, ts: number, callback: (error: Error | null, usage: number) => void): void
+  reserve (key: string, limit: number): Promise<number>
+  reserve (key: string, limit: number, callback: (error: Error | null, ts: number) => void): void
+
+  cancel (key: string, ts: number): Promise<number>
+  cancel (key: string, ts: number, callback: (error: Error | null, canceled: number) => void): void
 
   destroy (): void
 }